+++ src/variables/datasourceVariables.ts
@@ -8,6 +8,7 @@ export const ELASTIC_VARIABLE_NAME = 'elasticsearch';
 export const RQLITE_VARIABLE_NAME = 'infraDatasource';
 export const CLUSTER_VARIABLE_NAME = 'cluster';
 export const NAMESPACE_VARIABLE_NAME = 'namespace';
+export const WORKLOAD_VARIABLE_NAME = 'workload';
 
 export function createThanosDatasourceVariable() {
   return new DataSourceVariable({
@@ -59,6 +60,23 @@ export function createNamespaceFilterVariable(options: { isMulti?: boolean } = {
     query: {
       refId: 'namespaceVariableQuery',
       query: `label_values(kube_namespace_status_phase{cluster=~"\${${CLUSTER_VARIABLE_NAME}:regex}"}, namespace)`,
+    },
+    isMulti,
+    includeAll: isMulti,
+    allValue: isMulti ? '.+' : undefined,
+    value: isMulti ? '$__all' : '',
+  });
+}
+
+export function createWorkloadFilterVariable(options: { isMulti?: boolean } = {}) {
+  const isMulti = options.isMulti ?? true;
+  return new QueryVariable({
+    name: WORKLOAD_VARIABLE_NAME,
+    label: 'Workload',
+    datasource: { uid: `\${${THANOS_VARIABLE_NAME}}` },
+    query: {
+      refId: 'workloadVariableQuery',
+      query: `label_values(kube_pod_owner{cluster=~"\${${CLUSTER_VARIABLE_NAME}:regex}", namespace=~"\${${NAMESPACE_VARIABLE_NAME}:regex}"}, owner_name)`,
     },
     isMulti,
     includeAll: isMulti,